requirements.txt
python-telegram-bot
google-generativeai
pydantic
python-dotenv
sqlalchemy
gtts
